harsh-work/requirements.txt
torch
biopython
dendropy
